+++ assets/rings/rings-chasarooni;-runes-thin-ring-(black).json
@@ -1,0 +1,96 @@
+{
+
+	"config": {
+		"defaultColorBand": {
+			"startRadius": 0,
+			"endRadius": 1
+		},
+		"defaultRingColor": "#a51ee6"
+	},
+
+	"frames": {
+
+		"token-ring-gargantuan-bkg":
+		{
+			"frame": {"x":0,"y":0,"w":2048,"h":2048},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":2048,"h":2048},
+			"sourceSize": {"w":2048,"h":2048},
+			"anchor": {"x":0.5,"y":0.5}
+		},
+		"token-ring-gargantuan":
+		{
+			"frame": {"x":2048,"y":0,"w":2048,"h":2048},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":2048,"h":2048},
+			"sourceSize": {"w":2048,"h":2048},
+			"anchor": {"x":0.5,"y":0.5},
+			"gridTarget": 3
+		},
+		"token-ring-large-huge-bkg":
+		{
+			"frame": {"x":4096,"y":0,"w":1024,"h":1024},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":1024,"h":1024},
+			"sourceSize": {"w":1024,"h":1024},
+			"anchor": {"x":0.5,"y":0.5}
+		},
+		"token-ring-large-huge":
+		{
+			"frame": {"x":5120,"y":0,"w":1024,"h":1024},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":1024,"h":1024},
+			"sourceSize": {"w":1024,"h":1024},
+			"anchor": {"x":0.5,"y":0.5},
+			"gridTarget": 2
+		},
+		"token-ring-med-bkg":
+		{
+			"frame": {"x":6144,"y":0,"w":512,"h":512},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":512,"h":512},
+			"sourceSize": {"w":512,"h":512},
+			"anchor": {"x":0.5,"y":0.5}
+		},
+		"token-ring-med":
+		{
+			"frame": {"x":6656,"y":0,"w":512,"h":512},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":512,"h":512},
+			"sourceSize": {"w":512,"h":512},
+			"anchor": {"x":0.5,"y":0.5},
+			"gridTarget": 1
+		},
+		"token-ring-tiny-bkg":
+		{
+			"frame": {"x":7168,"y":0,"w":256,"h":256},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":256,"h":256},
+			"sourceSize": {"w":256,"h":256},
+			"anchor": {"x":0.5,"y":0.5}
+		},
+		"token-ring-tiny":
+		{
+			"frame": {"x":7424,"y":0,"w":256,"h":256},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {"x":0,"y":0,"w":256,"h":256},
+			"sourceSize": {"w":256,"h":256},
+			"anchor": {"x":0.5,"y":0.5},
+			"gridTarget": 0.5
+		}},
+	"meta": {
+		"version": "1.1",
+		"image": "rings-chasarooni;-runes-thin-ring-(black)",
+		"format": "RGBA8888",
+		"size": {"w":7680,"h":2048},
+		"scale": "1"
+	}
+}
