+++ scripts/authorList.js
@@ -1,0 +1,12 @@
+export const AUTHORS = [
+  // { name: "Token Tool", link: "https://www.rptools.net/toolbox/token-tool/" },
+  { name: "Joeundsoo", link: "https://primaryimmune.org/donate" },
+  { name: "Chasarooni", link: "https://ko-fi.com/Chasarooni" },
+  { name: "Redeux", link: "https://ko-fi.com/redeux" },
+  {
+    name: "Eldritch Dream",
+    link: "https://www.patreon.com/EldritchDreamCreations",
+  },
+  { name: "", link: "" },
+  { name: "", link: "" },
+];
